+++ examples/platform/silabs/zephyr/include/CommonDeviceCallbacks.h
@@ -1,6 +1,6 @@
 /*
  *
- *    Copyright (c) 2020-2023 Project CHIP Authors
+ *    Copyright (c) 2025 Project CHIP Authors
  *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,13 +15,6 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-
-/**
- * @file CommonDeviceCallbacks.h
- *
- * Common Implementations for the DeviceManager callbacks for all applications
- *
- **/
 
 #pragma once
