+++ examples/platform/silabs/zephyr/source/main.cpp
@@ -1,5 +1,5 @@
 /*
- *    Copyright (c) 2023-2024 Project CHIP Authors
+ *    Copyright (c) 2025 Project CHIP Authors
  *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
